linux-qcom-base_6.6.bb: Correct the number of GPIOs in gpio-ranges for QCS615
The UFS_RESET pin is expected to be wired to the reset pin of the
primary UFS memory, it's a general purpose output pin. Reorder it and
expose it as a gpio, so that the UFS driver can toggle it.

The QCS615 TLMM pin controller has GPIOs 0-122, so correct the
gpio-rangs to 124.

Change-Id: Ia5f5c648432759b61b069dd3665259f952c34090
Signed-off-by: Lijuan Gao <quic_lijuang@quicinc.com>