recipes-security: removed sshd and local_login from permissive list
with new yocto version we see /root not correctly getting labled.
This issue is with ostree and yocto version. To unblock we are
adding code that will re-create /root link which is
pointing to  /var/roothome and removing the  local_login and sshd
which had issue in landing to its own home directory ( in root login )
its  /root .

Change-Id: I040c9f410838d98c4cca77ebf9eebe81da73d3a8
Signed-off-by: Ravi Kumar Siddojigari <quic_rsiddoji@quicinc.com>