QCLINUX: conf/machine: Add 'deferred_probe_timeout=15' to KERNEL_CMDLINE_EXTR
On QCS615, it takes around 10 seconds from all Kernel built-in drivers
are registers and first dynamic Kernel module driver is registered. When
it's longer than 10 seconds, it trigger deferred timeout so Kernel give
up probe device with built-in driver while dependency is not ready. For
example GPU SMMU driver, this leads to GPU/display driver not work.
Relax the timeout to 15 seconds so devices can be stably probed.

Change-Id: If561e93ce2d40080ed5a235f2f42fe8de1dcb304
Signed-off-by: Xinzheng Long <quic_xinlon@quicinc.com>