linux-qcom-base_6.6.bb: clk: qcom: gdsc: Update retain_ff sequence and timeout for GDSC
The retain_ff bit should be updated for a GDSC when it is under SW
control and ON. The current sequence needs to be fixed as the GDSC
needs to update retention and is moved to HW control which does not
guarantee the GDSC to be in enabled state.

During the GDSC FSM state, the GDSC hardware waits for an ACK and the
timeout for the ACK is 2000us as per design requirements.

Change-Id: Ic90b6c80bf62003942815d1c8475695763dcdbbd
Signed-off-by: Taniya Das <quic_tdas@quicinc.com>
Signed-off-by: Dongfang Zhao <quic_doz@quicinc.com>
Signed-off-by: Ramya Dasineni <quic_rdasinen@quicinc.com>