UBOOT: qcs9100: Add QCOM RPMH Power Domain Driver Support
Added support for Qualcomm RPMH power domain driver, which is
responsible for managing power domains on Qualcomm SoCs.
This is a port of the upstream Linux changes sa8775p to U-Boot:
https://web.git.kernel.org/pub/scm/linux/kernel/git/torvalds/linux.git/tree/drivers/pmdomain/qcom/rpmhpd.c?id=3d25d46a255a83f94d7d4d4216f38aafc8e116b0
The driver file in U-Boot is in drivers/power/domain/qcom-rpmhpd.c.
The power domain driver currently has support to power
on and off MMCX power domain; Support for other power domains
can be added in this driver.

Change-Id: I1a9d9f3c4412ff8f728dbf146e47605b580bfbf8
Signed-off-by: Balaji Selvanathan <quic_bselvana@quicinc.com>
Signed-off-by: Ramya Dasineni <quic_rdasinen@quicinc.com>