UBOOT: qcs9100: Fix for UFS write issue
Encountered an issue with the MICRON Prod-MT128GAVAT2U31AA,
Rev-0103 device, where the initial SCSI write operation
consistently failed. Upon investigation, it was determined
that the problem caused from cache handling in the UFS device.

To address this, we disabled the write cache in the UFS by
setting the FUA bit in the second byte of the WRITE10 command.

Since there haven't been any issues with read cache handling
so far, cache was disabled in write operation alone.

Change-Id: Ibfa679b3cac6e2ba98793f2b28ced384b6cd12ef
Signed-off-by: Aswin Murugan <quic_aswinm@quicinc.com>
Signed-off-by: Ramya Dasineni <quic_rdasinen@quicinc.com>